requirements.txt
pysolr==3.5.0
PyYAML==3.11
requests==2.10.0
watson-developer-cloud==0.14.0
